+++ hsrb_quick_path_optimizer/test/joint_limit-test.cpp
@@ -25,7 +25,7 @@ LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT
 OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
 DAMAGE.
 */
-/// @brief A group of functions that acquire optimization parameters
+/// @brief Function group to obtain optimization parameters
 
 #include <string>
 #include <vector>