+++ hsrb_robot_local_planner_node/test/plugins_for_tests.cpp
@@ -86,7 +86,7 @@ class GeneratorPlugin : public tmc_robot_local_planner::IGenerator {
   bool Generate(const tmc_robot_local_planner::Constraints& constraints,
                 const tmc_manipulation_types::RobotState& initial_state,
                 double normalized_velocity,
-                const std::vector<std::string>& ignore_joints,
+                bool enable_base,
                 std::function<bool()> interrupt,
                 std::vector<tmc_manipulation_types::TimedRobotTrajectory>& trajectories_out) override {
     if (constraints.hard_joint_constraints.empty()) {
@@ -151,6 +151,7 @@ class ValidatorPlugin : public tmc_robot_local_planner::IValidator {
   }
 
   bool Validate(const std::vector<tmc_manipulation_types::TimedRobotTrajectory>& trajectories_in,
+                const tmc_manipulation_types::RobotState& initial_state,
                 std::function<bool()> interrupt,
                 tmc_manipulation_types::TimedRobotTrajectory& trajectory_out) override {
     if (trajectories_in.empty()) {