+++ hsrb_robot_local_planner_node/CHANGELOG.rst
@@ -2,6 +2,18 @@
 Changelog for package hsrb_robot_local_planner_node
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+2.0.0 (2025-12-04)
+-------------------
+* Add example launch file and command-sending script
+* When the remaining time is short, avoid planning a new trajectory
+* Fix an issue where a stop trajectory was not generated when the set of active joints changed
+* Adjust and implement interfaces to remove robot-specific dependencies.
+* Add functionality to notify when the input robot state is invalid.
+* Add error notification when odometry is not subscribed.
+* Update add imlc
+* Add goal relative linear constraint and soft path joint constraints
+* Contributors: Keisuke Takeshita, Tomohiro Ono
+
 2.0.0 (2024-10-16)
 -------------------
 * Initial release